# perf: add CowStr type to be able to use Cow::Borrowed on keys