--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,7 +14,7 @@ repository = "https://github.com/PSeitz/serde_json_borrow"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = "1.0.145"
+serde = { version = "1.0.145", features = ["derive"] }
 serde_json = "1.0.86"
 
 [dev-dependencies]