--- a/src/lib.rs
+++ b/src/lib.rs
@@ -78,6 +78,8 @@ mod object_vec;
 mod owned;
 mod ser;
 mod value;
+
+#[cfg(feature = "cowkeys")]
 mod cowstr;
 
 pub use object_vec::{KeyStrType, ObjectAsVec, ObjectAsVec as Map};