@PSeitz 2025-01-24
Sweet that's some big gains. It's even faster than simd json now